Added support for missing 'desc' attributes on boards and cards
The 'desc' attribute is optional with the Trello API for board and card
object, and may be missing from the returned JSON. Python raises a
KeyError when attempting to access a missing dictionary key
(http://wiki.python.org/moin/KeyError), but provides an alternative get
method that does not. Currently an empty string ('') is used as the
default description in the event of a missing 'desc' value.