Added fetch_actions() to the Board class
You can fetch the actions on an entire board, as well as just the lists or cards. (https://trello.com/docs/api/board/index.html#get-1-boards-board-id) This is useful if you want to snag all of a particular action on an entire board (rather then each card/list individually).