Add missing initials for members of an organization (fixes #176)
Unless a filter is supplied, the list of members for an organization
will only include the member full name and username.  This means that a
list of member objects created from an organization will lack data in the
'initials' field that would exist if the member object were created by id.