Fixes Issue #247 KeyError: 'customFieldItems'
The previous support for custom fields added to py-trello did not
add the parameter[1] required to retrieve custom field items on cards.
This resulted in an error when calling any method to reteive cards on
a given board.

This patch adds the support for retrieving custom field items when
retrieving collections of cards by adding the designated
"customFieldItems" parameter to the open_cards, closed_cards and
all_cards methods.  The value is defaulted to 'true', but allows
for override for the cases you don't need the customFieldItems.