Modifies Cards.from_json to not require customFieldItems
The Cards.from_json method expects that each time you query for
a card, the 'customFieldItems' key will be present in the json
returned from the API.  This causes an error if custom fields is
not enabled for the board or in the case where you're adding a
new card.

This patch changes the way the dictionary is accessed so that
when the customFieldItems key is not present it will return a
json collection with 0 length instead of raising a keyError.