use a session for http requests
this modifies py-trello to use a requests.Session object for fetching
data. This enables connection pooling and makes it possible to set
default parameters for all requests.  It also lets us expose the
Session object to the caller for fetching arbitrary Trello URLs (e.g.,
unsupported api methods, or board web urls, which allows us to extract
the board id by copy-and-pasting the board url from a browser).