Allow set_custom_field to accept an empty value
The Trello API allows clearing a custom field's value by sending an
empty PUT request (see
https://developer.atlassian.com/cloud/trello/guides/rest-api/getting-started-with-custom-fields/#clearing-customfielditems).
In keeping with that, this commit allows the card.set_custom_field
method to clear the given custom field when passed an empty string.
It's very simple - if the value is empty, it skips giving it a type and
just sends the empty JSON, causing the field in Trello to be unset.

In the future, having a dedicated clear_custom_field may make more
sense, but I decided to put the functionality in set_custom_field so as
to match the structure of the actual REST API more closely