fix #325: allow working with duplicate checklist item names
To allow dealing with duplicate checklist item names and to preserve
backwards compatibility, allow the argument of `delete_checklist_item`
to be either a checklist item object or a name string. If the name is
duplicate, pick the first. Additionally, change the `clear()` method
to use this functionality, to correctly clean checklists containing
items with duplicate names.