Allow board and list classes to be created from minimal json
When requesting resources you can specify only certain fields on those
resources are returned. This can be as minimal as only the IDs. This
means that when parsing json into resources, we can't assume any field
other than the id will be present.

All other model resources should also be updated, but only doing List
and Board for now.